#21 team 도메인 예외 처리 적용

+++ src/main/java/com/example/matchup/matchupbackend/controller/TeamController.java
@@ -1,10 +1,17 @@
 package com.example.matchup.matchupbackend.controller;
 
-import com.example.matchup.matchupbackend.dto.*;
 import com.example.matchup.matchupbackend.dto.mentoring.MentoringCardResponse;
+import com.example.matchup.matchupbackend.dto.request.team.TeamCreateRequest;
+import com.example.matchup.matchupbackend.dto.request.team.TeamSearchRequest;
+import com.example.matchup.matchupbackend.dto.response.team.MeetingSpot;
+import com.example.matchup.matchupbackend.dto.response.team.SliceTeamResponse;
+import com.example.matchup.matchupbackend.dto.response.team.TeamDetailResponse;
+import com.example.matchup.matchupbackend.dto.response.team.TeamType;
+import com.example.matchup.matchupbackend.error.exception.AuthorizeException;
 import com.example.matchup.matchupbackend.global.config.jwt.TokenProvider;
 import com.example.matchup.matchupbackend.service.TeamService;
 import io.swagger.v3.oas.annotations.Operation;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -27,27 +34,28 @@ public class TeamController {
     @GetMapping("/list/team")
     @ResponseStatus(HttpStatus.OK)
     @Operation(description = "메인 페이지 API")
-    public SliceTeamResponse showTeams(TeamSearchRequest teamSearchRequest, Pageable pageable) {
+    public SliceTeamResponse showTeams(@Valid TeamSearchRequest teamSearchRequest, Pageable pageable) {
         return teamService.searchSliceTeamResponseList(teamSearchRequest, pageable);
     }
 
     @PostMapping("/team")
     @ResponseStatus(HttpStatus.CREATED)
     @Operation(description = "팀 생성 및 저장")
-    public Long makeTeam(@RequestHeader(value = HEADER_AUTHORIZATION) String token, @RequestBody TeamCreateRequest teamCreateRequest) {
+    public Long makeTeam(@RequestHeader(value = "Authorization") String token, @Valid @RequestBody TeamCreateRequest teamCreateRequest) {
         Long userId = tokenProvider.getUserId(token);
-        return teamService.makeNewTeam(userId, teamCreateRequest);
+        if(userId == null){
+            throw new AuthorizeException("TeamCreate");
+        }
+            return teamService.makeNewTeam(userId, teamCreateRequest);
     }
 
     //팀 내용 업데이트
     @PutMapping("/team/{teamID}")
     @Operation(description = "팀 정보 수정") //인증 정보 추가돼서 팀장만 삭제할수 있도록 함
-    public ResponseEntity<String> updateTeam(@RequestHeader(value = HEADER_AUTHORIZATION) String token, @PathVariable Long teamID, @RequestBody TeamCreateRequest teamCreateRequest) {
+    public ResponseEntity<String> updateTeam(@RequestHeader(value = HEADER_AUTHORIZATION) String token, @PathVariable Long teamID, @Valid @RequestBody TeamCreateRequest teamCreateRequest) {
         Long userId = tokenProvider.getUserId(token);
-        if (teamService.isUpdatable(teamID, teamCreateRequest) == false) //업데이트 불가능
-        {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("기존 팀원 인원수 보다 높게 인원수를 설정하세요");
+        if (userId == null) {
+            throw new AuthorizeException("TeamUpdate");
         }
         teamService.updateTeam(userId, teamID, teamCreateRequest);
         return ResponseEntity.ok("업데이트 완료");
@@ -57,6 +65,9 @@ public class TeamController {
     @Operation(description = "팀 삭제") //인증 정보 추가돼서 팀장만 삭제할수 있도록 함
     public ResponseEntity<String> deleteTeam(@RequestHeader(value = HEADER_AUTHORIZATION) String token, @PathVariable Long teamID) {
         Long userId = tokenProvider.getUserId(token);
+        if (userId == null) {
+            throw new AuthorizeException("TeamDelete");
+        }
         teamService.deleteTeam(userId, teamID);
         return ResponseEntity.ok("팀 삭제 완료");
     }
@@ -91,7 +102,7 @@ public class TeamController {
 
     @GetMapping("/team/{teamID}/type")
     @ResponseStatus(HttpStatus.OK)
-    @Operation(description = "팀 상세페이지의 모임 카테고리")
+    @Operation(description = "팀 상세페이지의 모임 타입 ex) 프로젝트 - 웹프로젝트")
     public TeamType showTeamType(@PathVariable Long teamID) {
         return teamService.getTeamType(teamID);
     }
